refactor: move autotranslate modules
