--- apps/meteor/app/autotranslate/client/index.ts
@@ -1,3 +1,0 @@
-import './lib/actionButton';
-
-export { AutoTranslate, createAutoTranslateMessageStreamHandler } from './lib/autotranslate';
