--- apps/meteor/app/autotranslate/client/lib/actionButton.ts
@@ -1,3 +1,0 @@
-import { AutoTranslate } from './autotranslate';
-
-AutoTranslate.init();
