+++ apps/meteor/client/components/message/toolbar/useViewOriginalTranslationAction.ts
@@ -2,8 +2,8 @@ import type { IMessage, IRoom, ISubscription } from '@rocket.chat/core-typings';
 import { useEndpoint, usePermission, useSetting, useUser } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 
-import { AutoTranslate } from '../../../../app/autotranslate/client';
 import type { MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import { AutoTranslate } from '../../../lib/autotranslate';
 import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
 import { Messages } from '../../../stores';
 import { hasTranslationLanguageInAttachments, hasTranslationLanguageInMessage } from '../../../views/room/MessageList/lib/autoTranslate';
