+++ apps/meteor/client/importPackages.ts
@@ -1,4 +1,3 @@
-import '../app/autotranslate/client';
 import '../app/emoji/client';
 import '../app/gitlab/client';
 import '../app/license/client';
