+++ apps/meteor/client/lib/autotranslate/index.ts
@@ -1,0 +1,1 @@
+export { AutoTranslate, createAutoTranslateMessageStreamHandler } from './autotranslate';
