+++ apps/meteor/client/views/room/MessageList/hooks/useAutoTranslate.ts
@@ -2,7 +2,7 @@ import type { IMessage, ISubscription, ITranslatedMessage } from '@rocket.chat/c
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useCallback, useMemo } from 'react';
 
-import { AutoTranslate } from '../../../../../app/autotranslate/client';
+import { AutoTranslate } from '../../../../lib/autotranslate';
 import { roomCoordinator } from '../../../../lib/rooms/roomCoordinator';
 import { hasTranslationLanguageInAttachments, hasTranslationLanguageInMessage } from '../lib/autoTranslate';
 import { isOwnUserMessage } from '../lib/isOwnUserMessage';
